requirements.txt
astroid==2.0.4
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
dialogflow==0.5.1
Django==2.1.3
google-api-core==1.6.0
google-auth==1.6.1
googleapis-common-protos==1.5.5
grpcio==1.17.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.2
pandas==0.24.1
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.8.0
pytz==2018.7
requests==2.20.1
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
wcwidth==0.1.7
wit==5.1.0
wrapt==1.10.11
